Conflicts:
	episodes.list.d/season15.list

+++ bin/latest-hak5.sh
@@ -1,8 +1,13 @@
 #!/bin/bash
 # hak5 weekly
-dldir="/storage/video/podcasts/Hak5/S14"
+
+dldir="/storage/video/podcasts/Hak5/S15"
 sdir="/storage/video/podcasts/Hak5"
-wget -O $sdir/weeklyfeed "http://revision3.com/hak5/feed/MP4-hd30?subshow=false"
+
+wget -q -O $sdir/weeklyfeed "http://revision3.com/hak5/feed/MP4-hd30?subshow=false"
 filename=`cat $sdir/weeklyfeed | grep "mp4" | egrep -v '<guid|<atom|<itunes' | sed 's/^.*url="//' | sed 's/".*$//' | uniq | head -n1`
-wget -P $dldir -c $filename
-rm $sdir/weeklyfeed
+wget -q -P $dldir -c $filename
+
+if [ -f $sdir/weeklyfeed ]; then
+  rm $sdir/weeklyfeed
+fi
